platina-mk1_defconfig: Correct errors and shrink kernel for goes-boot
There were several configuration changes which slipped in as a result
of commit a6771f9bec. Restore these options to the proper value.

The base kernel had too many features built in to fit within the
bootrom for the goes-boot images. Make most features that are not
needed for booting be modules.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>